Refactored Search Result components for API Collection & Catalog
