+++ src/api/resources.js
@@ -1,5 +1,5 @@
 // Uniform Resource Locator : address of a given resource on the web
-const inProduction = false;
+const inProduction = true;
 
 export const api = {
   skryfallURL: 'https://api.scryfall.com',
