+++ src/components/views/search/CardImageSection.js
@@ -4,7 +4,8 @@ import ExpandIcon from './ExpandIcon';
 const CardImageSection = (props) => {
   const { attributes, setDisplay } = props;
 
-  const handleClick = () => {
+  const handleClick = (e) => {
+    console.log(e.target)
     setDisplay(true);
   }
