+++ src/css/form.css
@@ -401,7 +401,7 @@ input#search-collection {
     text-align: right;
 }
 
-.search-items {
+div.cards {
     padding-inline: 2rem;
 }
 
@@ -493,12 +493,6 @@ ul.autocomplete-list {
     display: flex;
     flex-direction: column;
     justify-content: space-between;
-  }
-
-  .card-wrapper {
-    display: flex;
-    flex-direction: column;
-    gap: 1rem;
   }
