+++ src/css/style.css
@@ -119,6 +119,10 @@ a {
     min-height: calc(100vh - var(--header-height));
 }
 
+.loading {
+  display: block;
+}
+
 .spinner p {
     padding-block-start: 5rem;
 }
